HelloWorld/requirements.txt
awsiotsdk
click
